refactor(queries) : reorganize query files and update imports
