+++ src/app/sn/users/[username]/UserContent/Cards/CardMenu/RecipeEdit/index.tsx
@@ -6,8 +6,8 @@ import style from './style.module.scss';
 
 import { NewRecipeDataServer } from '@/services/recipes/type';
 
-import { recipeOptions } from '@/queries/recipeOptions';
-import useEditRecipeMutation from '@/queries/hooks/useEditRecipeMutation';
+import { useEditRecipeMutation } from '@/queries/hooks';
+import { recipeOptions } from '@/queries/options';
 
 import Skeleton from '@/components/Skeleton';
