+++ src/app/sn/users/[username]/UserContent/RecipeList.tsx
@@ -3,7 +3,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { notFound, useParams } from 'next/navigation';
 
-import { recipeListOptions } from '@/queries/recipeListOptions';
+import { recipeListOptions } from '@/queries/options';
 
 import { useClientStore } from '@/providers/client-store-provider';
